+++ nylas/client/restful_models.py
@@ -20,7 +20,14 @@ def typed_dict_attr(items, attr_name=None):
     return dct
 
 
-class NylasAPIObject(dict):
+def _is_subclass(cls, parent):
+    for base in cls.__bases__:
+        if base.__name__.lower() == parent:
+            return True
+    return False
+
+
+class RestfulModel(dict):
     attrs = []
     date_attrs = {}
     datetime_attrs = {}
@@ -36,7 +43,7 @@ class NylasAPIObject(dict):
         self.id = None
         self.cls = cls
         self.api = api
-        super(NylasAPIObject, self).__init__()
+        super(RestfulModel, self).__init__()
 
     __setattr__ = dict.__setitem__
     __delattr__ = dict.__delitem__
@@ -46,7 +53,12 @@ class NylasAPIObject(dict):
     def create(cls, api, **kwargs):
         object_type = kwargs.get("object")
         cls_object_type = getattr(cls, "object_type", cls.__name__.lower())
-        if object_type and object_type != cls_object_type and object_type != "account":
+        if (
+            object_type
+            and object_type != cls_object_type
+            and object_type != "account"
+            and not _is_subclass(cls, object_type)
+        ):
             # We were given a specific object type and we're trying to
             # instantiate something different; abort. (Relevant for folders
             # and labels API.)
@@ -111,6 +123,11 @@ class NylasAPIObject(dict):
                     for value in values:
                         dct[attr].append(value)
         return dct
+
+
+class NylasAPIObject(RestfulModel):
+    def __init__(self, cls, api):
+        RestfulModel.__init__(self, cls, api)
 
     def child_collection(self, cls, **filters):
         return RestfulModelCollection(cls, self.api, **filters)